End of section 10

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
 using System;
+using Application.Core;
 using MediatR;
 using Persistence;
 
@@ -6,21 +7,31 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-  public class Command : IRequest
+  public class Command : IRequest<Result<Unit>>
   {
     public required string Id { get; init; }
   }
 
-  public class Handler(AppDbContext context) : IRequestHandler<Command>
+  public class Handler(AppDbContext context) : IRequestHandler<Command, Result<Unit>> 
   {
-    public async Task Handle(Command request, CancellationToken cancellationToken)
+    public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
     {
-      var activity = await context.Activities.FindAsync([request.Id], cancellationToken: cancellationToken)
-        ?? throw new ArgumentException("Cannot find activity");
+      var activity = await context.Activities.FindAsync([request.Id], cancellationToken: cancellationToken);
+
+      if (activity is null)
+      {
+        return Result<Unit>.Failure("Activity not found", 404);
+      }
 
       context.Activities.Remove(activity);
+      var isSuccess = (await context.SaveChangesAsync(cancellationToken)) > 0;
 
-      await context.SaveChangesAsync(cancellationToken);  
+      if(!isSuccess)
+      {
+        return Result<Unit>.Failure("Failed to delete the activity", 400);
+      }
+
+      return Result<Unit>.Success(Unit.Value);
     }
   }
 }
